+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 # Changelog
 
+Copyright (C) 2026 gitzip contributors
+
+This file is part of gitzip.
+
+gitzip is free software: you can redistribute it and/or modify it under
+the terms of the GNU General Public License as published by the Free Software
+Foundation, either version 3 of the License, or (at your option) any later
+version.
+
 All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, with a simple versioned history for this repository.
 
 ## [Unreleased]
 
-- No unreleased changes yet.
+### Changed
+
+- Switched the project license from Apache 2.0 to GPLv3 or later and updated repository notices
 
 ## [v0.0.1] - 2026-04-17